+++ src/xmlrpc/nightreq.h
@@ -36,11 +36,11 @@ class Night: public GetRequestAuthorized
 	public:
 		Night (const char *prefix, XmlRpc::XmlRpcServer *s):GetRequestAuthorized (prefix, s) {};
 
-		virtual void authorizedExecute (std::string path, XmlRpc::HttpParams *params, const char* &response_type, char* &response, int &response_length);
+		virtual void authorizedExecute (std::string path, XmlRpc::HttpParams *params, const char* &response_type, char* &response, size_t &response_length);
 	private:
-		void printObs (int obs_id, const char* &response_type, char* &response, int &response_length);
+		void printObs (int obs_id, const char* &response_type, char* &response, size_t &response_length);
 		void listObs (int year, int month, int day, std::ostringstream &_os);
-		void printTable (int year, int month, int day, char* &response, int &response_length);
+		void printTable (int year, int month, int day, char* &response, size_t &response_length);
 
 };
